% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.SNODAS.R
\name{download.SNODAS}
\alias{download.SNODAS}
\title{Downloads SNODAS data from the UC Boulder FTP server}
\arguments{
\item{dates}{A vector of date objects or strings that can be converted to dates by lubridate::ymd()}

\item{path}{A path to save the files locally, with a trailing slash (e.g. the default is './SNODAS/'). If the directory does not exist it will be created.}

\item{overwrite}{Should extant files be overwritten?}

\item{masked}{Do you want CONUS snodas data or unmasked North America (NB: unmasked has much more restricted dates available)}

\item{parallel}{Do you want to download and process files in parallel (for greatly enhanced performance?) via the doParallel package.}

\item{ncores}{If parallel=TRUE, how many cores should R use for the process. Default is 3 as a "safe" number, but can be adjusted higher if your computer has more cores available.}

\item{unzip}{Should the files be unzipped or will they be accessed in their compressed state? Unzipping dramatically increases read performance but will take up much more hard drive space. If you're only going to make one spatial subset of the data, leaving the files comopressed will be faster and take up less storage. But if you're going to use the SNODAS data over and over again, unzip the data (storage space permitting)}
}
\value{
NULL (This function is only used for its side effects. SNODAS data for the specified dates will be downloaded and untarballed in the specified path.)
}
\description{
Downloads SNODAS data from the UC Boulder FTP server
}
\examples{
download.SNODAS(dates=c('2010-02-20', '2015-02-10'), path='./SNODAStest/')

}
